# Conflicts:
#   BE/climbingBear/src/main/java/com/example/climbingBear/domain/mntn/dto/MntnDetailResDto.java
#   BE/climbingBear/src/main/java/com/example/climbingBear/domain/mntn/dto/MntnListResDto.java
#   BE/climbingBear/src/main/java/com/example/climbingBear/domain/mntn/entity/Mountain.java
#   BE/climbingBear/src/main/java/com/example/climbingBear/domain/record/repository/RecordRepository.java
#   BE/climbingBear/src/main/java/com/example/climbingBear/domain/record/service/RecordService.java
#   BE/climbingBear/src/main/java/com/example/climbingBear/domain/user/service/UserService.java
#   BE/climbingBear/src/main/resources/application.properties

+++ BE/climbingBear/src/main/java/com/example/climbingBear/domain/mntn/entity/Spot.java
@@ -23,8 +23,8 @@ public class Spot {
     private String spotNum;
 
     @Column(name = "lat")
-    private Float lat;
+    private double lat;
 
     @Column(name = "lon")
-    private Float lon;
+    private double lon;
 }
